penambahan diagram lingkaran di halaman beranda
