+++ src/route/MainRoute.jsx
@@ -4,8 +4,9 @@ import { store } from '../store/store';
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
 import BerandaAdmin from '../pages/Beranda';
 import Masuk from '../pages/Masuk';
-import Pengguna from '../pages/Pengguna';
+import Keluhan from '../pages/Keluhan';
 import DetailKeluhan from '../pages/DetailKeluhan';
+import Pengguna from '../pages/Pengguna';
 import Komentar from '../pages/Komentar';
 
 const MainRoute = () => {
@@ -15,8 +16,9 @@ const MainRoute = () => {
                 <Switch>
                     <Route exact path="/" component={BerandaAdmin} />
                     <Route exact path='/masuk' component={Masuk} />
-                    <Route exact path='/pengguna' component={Pengguna} />
+                    <Route exact path="/keluhan" component={Keluhan} />
                     <Route exact path='/keluhan/:id' component={DetailKeluhan} />
+                    <Route exact path='/pengguna' component={Pengguna} />
                     <Route exact path='/komentar' component={Komentar} />
                 </Switch>
             </BrowserRouter>
